refactor(mitm): extract provisionDnsEntries to dns/provision.ts (manager.ts under file-size cap) (#6127)
